move verification confirmation process to the verification model and controller
Before, confirming a verification was being handled as an update action on a subscription model. Now the form the enter is rendered through VerificationsController:edit and patches to VerificationsController:update. All the appropriate routes. Also the added benenfit of not having to rely on callbacks in the subscription model, making subscription instances much easier to work with.