better validate attendance_record to unique user_id per day
The old validation required instantiating the associated user model,
then all of its other attendance records and making sure the created_at
doesn't match the current day. This created a lot of overhead for what
should be a simple query. The new way uses a condition to scope the
uniqueness check on user_id to just records for today.