changed the way flash messages are rendered
We needed to add an html_safe segment to the flash message. The best
solution was to have the flash helper accept the block, in which we
pass the flashes we want to render. This allows us to call html_safe
on the parts of the flash we want to be html safe.