adjust relationships surrounding assessments
Before grades served as a join between an assessment and a requirement
to give a score for the one requirement on a submission. This groups
grades together into a review, allowing us to have a note that isn't
tied to a specific grade, as well as a user who did the reviewing.