fix test conflicts with number of absences
Calculating the number of absences is dependent on knowing how
many days have passed since the start of class. The original thought
was to only count days that have passed. This made calculating
attendances not work as planned. After further consideration, it
did make more sense to count the current day as well as the ones
that have passed.