adjust cohort factories to have a start_date on monday and end_date on friday
Running the tests near or on the weekend caused cohort related specs to fail.
This was because the start date of the factory cohort was dependent on the
current date. Now it's still based on the current date, but always falls to
the monday (for start_date) or friday (for end_date) of the week.