add score model
Previously the Grade model held a score. The problem was associating
a score with a description (which is what we want to show to the user).
The Score object holds both a value and a description and can be
manually created/edited if we ever want to change the score options.