prepopulate new review form with information from previous review
When a review is created, we wanted to be able to see what feedback
had been given in the previous review. We didn't, however, want to
simply update the old review. Therefore, when a review is created,
if there is a previous review, we build it off that review using
the deep_cloneable gem.