fix attendance_record uniqueness validation
In an attempt to refactor the original solution for this, we instead
broke it. We didn't know because we didn't test it after 5. I am
officially deeming this the gremlin bug, because it only show up
after midnight (utc). This commit kills the gremlin, so all is
right in the world.