switch attendance statistics show action to index
This was originally a show action because we were displaying a single
model (CohortAttendanceStatistics). Now we want to add the ability
for a student to see their individual attendance statistics. Since
this deals with multiple attendance records, we thought it made sense to
move it to the index action to allow the show action to be for an
individual student's attendance stats.