create transcripts for students
This includes a list of assessments they have met expectations on.
Methods had to be added to assessments, submissions, and reviews
in order to simplify determining if student has met expecations on
an assessment or not. This is why you'll see Review#meets_expectations?
and Submission#meets_expectations? Otherwise the logic was getting
much to hairy.