add date property to attendance record
We're looking to add the ability to update attendance records manually. This mean we're looking at having attendance records created on a different day then the one it's supposed to represent. Therefore, we need to decouple an attendance record's date from the created_at field and dynamically set it before validation. This also makes dealing with validating an attendance records uniqueness by date much more straightforward.
date is was actually created on.