remove unnecessary method
When this was first written, the hope was to have separate methods for each of the three possible outcomes (creating a new attendance record, updating an existing one with the right properties, or destroying an existing one). destroy_attendance_record was the first method built with this in mind, but it's something that I soon moved away from. However, it was mistakenly left in. This commit removes that method.