refactor cohort model
Looking forward, I want to be able to give students feedback about their attendance record: what days they have missed or been tardy. This seems like it will be especially tricky for absences, since an absence is simple a lack of attendance record on a given day. In order to make this easier, cohorts need to be able to give you a collection of date objects. It works out nice, as cohorts were already filtering days. This separates out the filtering from the counting, making it more flexible.