actually rename class_dates to class_dates_until
Once there was a commit that declared the reworking of class_dates into
class_dates_until. But, alas, this was a lie. While changes were made to
class_dates, its name remained the same. But take heart! This commit
makes good on that promise, and actually changes the name of class_dates
to class_dates_until. And there was much rejoicing.